Move BaseFormatter into /lib/formatters
