+++ lib/formatters/base.js
@@ -22,7 +22,7 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-import * as exec from './exec';
+import * as exec from '../exec';
 
 export class BaseFormatter {
     constructor(formatterInfo) {
