+++ lib/formatters/gofmt.js
@@ -22,8 +22,9 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-import { BaseFormatter } from '../base-formatter';
 import * as exec from '../exec';
+
+import { BaseFormatter } from './base';
 
 export class GoFmtFormatter extends BaseFormatter {
     static get key() { return 'gofmt'; }
