+++ lib/formatters/index.js
@@ -26,7 +26,7 @@ import { makeKeyedTypeGetter } from '../keyed-type';
 
 import * as all from './_all';
 
-export { BaseFormatter } from '../base-formatter';
+export { BaseFormatter } from './base';
 export * from './_all';
 
 export const getFormatterTypeByKey = makeKeyedTypeGetter('formatter', all);
